requirements.txt
wheel==0.29.0
Django==1.9
django-appconf==1.0.1
django-compressor==1.6
django-dual-authentication==1.0.0
djangorestframework==3.3.2
djrill==2.0
mysqlclient==1.3.7
Pillow==3.0.0
requests==2.9.1
simplejson==3.8.1
six==1.10.0
yara==1.7.7
yara-python==3.4.0.0
sendgrid==2.2.1
sendgrid-django==2.0.0
python-magic==0.4.12